Move Run<X>Mode in dedicated files
