+++ src/tests/end-to-end/simple_study/CMakeLists.txt
@@ -3,31 +3,30 @@ enable_testing()
 
 #bigobj support needed for windows compilation
 if(MSVC)
-	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /bigobj")
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /bigobj")
 endif(MSVC)
 
 add_executable(tests-simple-study
-		simple-study.cpp
-		)
+        simple-study.cpp
+)
 
 target_link_libraries(tests-simple-study
-						PRIVATE
-						test_utils
-						antares-solver-main-economy
-						antares-solver-main-adequacy
-						antares-solver-hydro
-						antares-solver-variable
-						antares-solver-simulation
-						antares-solver-ts-generator
-						model_antares
-						${Boost_UNIT_TEST_FRAMEWORK_LIBRARY}
-			)
-			
+        PRIVATE
+        test_utils
+        Antares::run-mode
+        antares-solver-hydro
+        antares-solver-variable
+        antares-solver-simulation
+        antares-solver-ts-generator
+        model_antares
+        ${Boost_UNIT_TEST_FRAMEWORK_LIBRARY}
+)
+
 target_include_directories(tests-simple-study
-							PRIVATE
-							${CMAKE_SOURCE_DIR}/solver
-							${CMAKE_CURRENT_SOURCE_DIR}/../utils
-							)
+        PRIVATE
+        ${CMAKE_SOURCE_DIR}/solver
+        ${CMAKE_CURRENT_SOURCE_DIR}/../utils
+)
 
 add_test(NAME end-to-end-simple-study COMMAND tests-simple-study)
 set_property(TEST end-to-end-simple-study PROPERTY LABELS end-to-end)
